+++ src/general/abstract_system.jl
@@ -1,28 +1,20 @@
 # Abstract supertype for all system types.
-abstract type System{NDIMS} end
+abstract type AbstractSystem{NDIMS} end
 
-abstract type FluidSystem{NDIMS} <: System{NDIMS} end
-timer_name(::FluidSystem) = "fluid"
-vtkname(system::FluidSystem) = "fluid"
+@inline Base.ndims(::AbstractSystem{NDIMS}) where {NDIMS} = NDIMS
+@inline Base.eltype(system::AbstractSystem) = error("eltype not implemented for system $system")
 
-abstract type SolidSystem{NDIMS} <: System{NDIMS} end
-timer_name(::SolidSystem) = "solid"
-vtkname(system::SolidSystem) = "solid"
+abstract type AbstractFluidSystem{NDIMS} <: AbstractSystem{NDIMS} end
+timer_name(::AbstractFluidSystem) = "fluid"
+vtkname(system::AbstractFluidSystem) = "fluid"
 
-abstract type BoundarySystem{NDIMS} <: System{NDIMS} end
-timer_name(::BoundarySystem) = "boundary"
-vtkname(system::BoundarySystem) = "boundary"
+abstract type AbstractStructureSystem{NDIMS} <: AbstractSystem{NDIMS} end
+timer_name(::AbstractStructureSystem) = "structure"
+vtkname(system::AbstractStructureSystem) = "structure"
 
-@inline function set_zero!(du)
-    du .= zero(eltype(du))
-
-    return du
-end
-
-initialize!(system, semi) = system
-
-@inline Base.ndims(::System{NDIMS}) where {NDIMS} = NDIMS
-@inline Base.eltype(system::System) = error("eltype not implemented for system $system")
+abstract type AbstractBoundarySystem{NDIMS} <: AbstractSystem{NDIMS} end
+timer_name(::AbstractBoundarySystem) = "boundary"
+vtkname(system::AbstractBoundarySystem) = "boundary"
 
 # Number of integrated variables in the first component of the ODE system (coordinates)
 @inline u_nvariables(system) = ndims(system)
@@ -35,20 +27,30 @@ initialize!(system, semi) = system
 @inline nparticles(system) = length(system.mass)
 
 # Number of particles in the system whose positions are to be integrated (corresponds to the size of u and du)
-@inline n_moving_particles(system) = nparticles(system)
+@inline n_integrated_particles(system) = nparticles(system)
 
-@inline eachparticle(system::System) = active_particles(system)
+@inline eachparticle(system::AbstractSystem) = each_active_particle(system)
 @inline eachparticle(initial_condition) = Base.OneTo(nparticles(initial_condition))
 
 # Wrapper for systems with `SystemBuffer`
-@inline each_moving_particle(system) = each_moving_particle(system, buffer(system))
-@inline each_moving_particle(system, ::Nothing) = Base.OneTo(n_moving_particles(system))
+@inline each_integrated_particle(system) = each_integrated_particle(system, buffer(system))
+@inline function each_integrated_particle(system, ::Nothing)
+    return Base.OneTo(n_integrated_particles(system))
+end
 
 @inline active_coordinates(u, system) = active_coordinates(u, system, buffer(system))
 @inline active_coordinates(u, system, ::Nothing) = current_coordinates(u, system)
 
-@inline active_particles(system) = active_particles(system, buffer(system))
-@inline active_particles(system, ::Nothing) = Base.OneTo(nparticles(system))
+@inline each_active_particle(system) = each_active_particle(system, buffer(system))
+@inline each_active_particle(system, ::Nothing) = Base.OneTo(nparticles(system))
+
+@inline function set_zero!(du)
+    du .= zero(eltype(du))
+
+    return du
+end
+
+initialize!(system, semi) = system
 
 # This should not be dispatched by system type. We always expect to get a column of `A`.
 @propagate_inbounds function extract_svector(A, system, i)
@@ -100,11 +102,11 @@ end
 # By default, try to extract it from `v`.
 @inline current_velocity(v, system) = v
 
-@inline function current_density(v, system::System, particle)
+@inline function current_density(v, system::AbstractSystem, particle)
     return current_density(v, system)[particle]
 end
 
-@propagate_inbounds function current_pressure(v, system::System, particle)
+@propagate_inbounds function current_pressure(v, system::AbstractSystem, particle)
     return current_pressure(v, system)[particle]
 end