+++ src/schemes/fluid/weakly_compressible_sph/system.jl
@@ -30,7 +30,8 @@ See [Weakly Compressible SPH](@ref wcsph) for more details on the method.
 - `acceleration`:               Acceleration vector for the system. (default: zero vector)
 - `viscosity`:                  Viscosity model for this system (default: no viscosity).
                                 See [`ArtificialViscosityMonaghan`](@ref) or [`ViscosityAdami`](@ref).
-- `density_diffusion`:          Density diffusion terms for this system. See [`DensityDiffusion`](@ref).
+- `density_diffusion`:          Density diffusion terms for this system.
+                                See [`AbstractDensityDiffusion`](@ref TrixiParticles.AbstractDensityDiffusion).
 - `pressure_acceleration`:      Pressure acceleration formulation for this system.
                                 By default, the correct formulation is chosen based on the
                                 density calculator and the correction method.
@@ -40,7 +41,7 @@ See [Weakly Compressible SPH](@ref wcsph) for more details on the method.
                                 formulation](@ref transport_velocity_formulation) to use
                                 with this system. Default is no shifting.
 - `buffer_size`:                Number of buffer particles.
-                                This is needed when simulating with [`OpenBoundarySPHSystem`](@ref).
+                                This is needed when simulating with [`OpenBoundarySystem`](@ref).
 - `correction`:                 Correction method used for this system. (default: no correction, see [Corrections](@ref corrections))
 - `source_terms`:               Additional source terms for this system. Has to be either `nothing`
                                 (by default), or a function of `(coords, velocity, density, pressure, t)`
@@ -60,7 +61,8 @@ See [Weakly Compressible SPH](@ref wcsph) for more details on the method.
 - `color_value`:                The value used to initialize the color of particles in the system.
 """
 struct WeaklyCompressibleSPHSystem{NDIMS, ELTYPE <: Real, IC, MA, P, DC, SE, K, V, DD, COR,
-                                   PF, SC, ST, B, SRFT, SRFN, PR, C} <: FluidSystem{NDIMS}
+                                   PF, SC, ST, B, SRFT, SRFN, PR,
+                                   C} <: AbstractFluidSystem{NDIMS}
     initial_condition                 :: IC
     mass                              :: MA     # Array{ELTYPE, 1}
     pressure                          :: P      # Array{ELTYPE, 1}
@@ -410,7 +412,7 @@ function write_v0!(v0, system::WeaklyCompressibleSPHSystem, ::ContinuityDensity)
 end
 
 function restart_with!(system::WeaklyCompressibleSPHSystem, v, u)
-    for particle in each_moving_particle(system)
+    for particle in each_integrated_particle(system)
         system.initial_condition.coordinates[:, particle] .= u[:, particle]
         system.initial_condition.velocity[:, particle] .= v[1:ndims(system), particle]
     end
@@ -423,7 +425,7 @@ function restart_with!(system, ::SummationDensity, v, u)
 end
 
 function restart_with!(system, ::ContinuityDensity, v, u)
-    for particle in each_moving_particle(system)
+    for particle in each_integrated_particle(system)
         system.initial_condition.density[particle] = v[end, particle]
     end
 
@@ -434,7 +436,7 @@ end
     extract_smatrix(system.cache.correction_matrix, system, particle)
 end
 
-@inline function curvature(particle_system::FluidSystem, particle)
+@inline function curvature(particle_system::AbstractFluidSystem, particle)
     (; cache) = particle_system
     return cache.curvature[particle]
 end