+++ test/examples/examples.jl
@@ -3,19 +3,19 @@
 @testset verbose=true "Examples" begin
     include("examples_fluid.jl")
 
-    @testset verbose=true "Solid" begin
-        @trixi_testset "solid/oscillating_beam_2d.jl" begin
+    @testset verbose=true "Structure" begin
+        @trixi_testset "structure/oscillating_beam_2d.jl" begin
             @trixi_test_nowarn trixi_include(@__MODULE__,
-                                             joinpath(examples_dir(), "solid",
+                                             joinpath(examples_dir(), "structure",
                                                       "oscillating_beam_2d.jl"),
                                              tspan=(0.0, 0.1))
             @test sol.retcode == ReturnCode.Success
             @test count_rhs_allocations(sol, semi) == 0
         end
 
-        @trixi_testset "solid/oscillating_beam_2d.jl with penalty force and viscosity" begin
+        @trixi_testset "structure/oscillating_beam_2d.jl with penalty force and viscosity" begin
             @trixi_test_nowarn trixi_include(@__MODULE__,
-                                             joinpath(examples_dir(), "solid",
+                                             joinpath(examples_dir(), "structure",
                                                       "oscillating_beam_2d.jl"),
                                              tspan=(0.0, 0.1),
                                              penalty_force=PenaltyForceGanzenmueller(alpha=0.1),