+++ test/schemes/boundary/dummy_particles/rhs.jl
@@ -47,8 +47,8 @@
                                                                  PressureZeroing(),
                                                                  smoothing_kernel,
                                                                  smoothing_length)
-            boundary_system_zeroing = BoundarySPHSystem(initial_condition,
-                                                        boundary_model_zeroing)
+            boundary_system_zeroing = WallBoundarySystem(initial_condition,
+                                                         boundary_model_zeroing)
             boundary_model_continuity = BoundaryModelDummyParticles(initial_condition.density,
                                                                     initial_condition.mass,
                                                                     ContinuityDensity(),
@@ -56,8 +56,8 @@
                                                                     smoothing_length)
             # Overwrite `boundary_model_continuity.pressure` because we skip the update step
             boundary_model_continuity.pressure .= initial_condition.pressure
-            boundary_system_continuity = BoundarySPHSystem(initial_condition,
-                                                           boundary_model_continuity)
+            boundary_system_continuity = WallBoundarySystem(initial_condition,
+                                                            boundary_model_continuity)
 
             boundary_model_summation = BoundaryModelDummyParticles(initial_condition.density,
                                                                    initial_condition.mass,
@@ -68,8 +68,8 @@
             boundary_model_summation.pressure .= initial_condition.pressure
             # Density is stored in the cache
             boundary_model_summation.cache.density .= initial_condition.density
-            boundary_system_summation = BoundarySPHSystem(initial_condition,
-                                                          boundary_model_summation)
+            boundary_system_summation = WallBoundarySystem(initial_condition,
+                                                           boundary_model_summation)
 
             u_boundary = zeros(0, TrixiParticles.nparticles(initial_condition))
             v_boundary = zeros(0, TrixiParticles.nparticles(initial_condition))
@@ -77,20 +77,20 @@
             v_boundary_continuity = copy(initial_condition.density')
 
             # TLSPH system
-            solid_system = TotalLagrangianSPHSystem(initial_condition, smoothing_kernel,
-                                                    smoothing_length, 0.0, 0.0,
-                                                    boundary_model=boundary_model_continuity)
+            structure_system = TotalLagrangianSPHSystem(initial_condition, smoothing_kernel,
+                                                        smoothing_length, 0.0, 0.0,
+                                                        boundary_model=boundary_model_continuity)
 
-            # Positions of the solid particles are not used here
-            u_solid = zeros(0, TrixiParticles.nparticles(solid_system))
-            v_solid = vcat(initial_condition.velocity,
-                           initial_condition.density')
+            # Positions of the structure particles are not used here
+            u_structure = zeros(0, TrixiParticles.nparticles(structure_system))
+            v_structure = vcat(initial_condition.velocity,
+                               initial_condition.density')
 
             systems = Dict(
                 "Fluid-Fluid" => second_fluid_system,
                 "Fluid-BoundaryDummyPressureZeroing" => boundary_system_zeroing,
                 "Fluid-BoundaryDummyContinuityDensity" => boundary_system_continuity,
-                "Fluid-TLSPH" => solid_system
+                "Fluid-TLSPH" => structure_system
             )
 
             if density_calculator isa SummationDensity
@@ -105,7 +105,7 @@
                 "Fluid-BoundaryDummyContinuityDensity" => (v_boundary_continuity,
                                                            u_boundary),
                 "Fluid-BoundaryDummySummationDensity" => (v_boundary, u_boundary),
-                "Fluid-TLSPH" => (v_solid, u_solid)
+                "Fluid-TLSPH" => (v_structure, u_structure)
             )
 
             return systems, vu