+++ NEWS.md
@@ -8,8 +8,21 @@ used in the Julia ecosystem. Notable changes will be documented in this file for
 
 ### API Changes
 
-- API for `OpenBoundarySPHSystem` and `BoundaryZone` changed.
-  It is now possible to pass multiple `BoundaryZone`s to a single `OpenBoundarySPHSystem`.
+- Renamed `BoundarySPHSystem` to `WallBoundarySystem` and the keyword argument
+  `movement` to `prescribed_motion`.
+
+- Renamed `OpenBoundarySPHSystem` to `OpenBoundarySystem`.
+
+- Renamed `BoundaryMovement` to `PrescribedMotion`.
+
+- Renamed directory `solid` to `structure` in the examples file tree.
+  VTK files for the `TotalLagrangianSPHSystem` are now also called `structure_*`.
+
+- Renamed keyword argument `n_fixed_particles` of the `TotalLagrangianSPHSystem`
+  to `n_clamped_particles`.
+
+- API for `OpenBoundarySystem` and `BoundaryZone` changed.
+  It is now possible to pass multiple `BoundaryZone`s to a single `OpenBoundarySystem`.
   Reference values are now assigned individually to each `BoundaryZone`. (#866)
 
 - The argument of `TransportVelocityAdami` is now a keyword argument.
@@ -29,6 +42,10 @@ used in the Julia ecosystem. Notable changes will be documented in this file for
   that were not using the documented API but were functions of
   `(system, v_ode, u_ode, semi, t)` now need to be functions
   of `(system, dv_ode, du_ode, v_ode, u_ode, semi, t)` (#879).
+
+- Renamed `each_moving_particle` to `each_integrated_particle`,
+  `n_moving_particles` to `n_integrated_particles`
+  and `active_particles` to `each_active_particle`.
 
 ### Features